src/main/java/local/mateo/cleanArchitecture/infraestructure/health/LivenessCheckSimple.java
package local.mateo.cleanArchitecture.infraestructure.health;

import org.eclipse.microprofile.health.HealthCheck;
import org.eclipse.microprofile.health.HealthCheckResponse;
import org.eclipse.microprofile.health.HealthCheckResponseBuilder;
import org.eclipse.microprofile.health.Liveness;

import jakarta.enterprise.context.ApplicationScoped;

@Liveness
@ApplicationScoped 
public class LivenessCheckSimple implements HealthCheck {

    @Override
    public HealthCheckResponse call() {
        HealthCheckResponseBuilder responseBuilder = HealthCheckResponse.named("Liveness: Simple health check");
        responseBuilder.up();
        return responseBuilder.build();
    }
}